j,yjdbk rjyabub

+++ rofi/powermenu.sh
@@ -1,42 +1,29 @@
-#!/usr/bin/env bash
-
-## Author : Aditya Shakya (adi1090x)
-## Github : @adi1090x
-#
-## Rofi   : Power Menu
-#
-## Available Styles
-#
-## style-1   style-2   style-3   style-4   style-5
-## style-6   style-7   style-8   style-9   style-10
-
-# Current Theme
-set dir "$HOME/.config/rofi"
-set theme 'powermenu'
+dir="$HOME/.config/rofi"
+theme='powermenu'
 
 # CMDs
-set uptime "`uptime -p | sed -e 's/up //g'`"
-set host `hostname`
+uptime="`uptime -p | sed -e 's/up //g'`"
+host=`hostname`
 
 # Options
-set shutdown ''
-set reboot ''
-set lock ''
-set suspend ''
-set logout ''
-set yes ''
-set no ''
+shutdown=''
+reboot=''
+lock=''
+suspend=''
+logout=''
+yes=''
+no=''
 
 # Rofi CMD
-function rofi_cmd
+rofi_cmd() {
 	rofi -dmenu \
 		-p "Uptime: $uptime" \
 		-mesg "Uptime: $uptime" \
-		-theme $dir/$theme.rasi
-end
+		-theme ${dir}/${theme}.rasi
+}
 
 # Confirmation CMD
-function confirm_cmd
+confirm_cmd() {
 	rofi -theme-str 'window {location: center; anchor: center; fullscreen: false; width: 350px;}' \
 		-theme-str 'mainbox {children: [ "message", "listview" ];}' \
 		-theme-str 'listview {columns: 2; lines: 1;}' \
@@ -45,38 +32,54 @@ function confirm_cmd
 		-dmenu \
 		-p 'Confirmation' \
 		-mesg 'Are you Sure?' \
-		-theme $dir/$theme.rasi
-end
+		-theme ${dir}/${theme}.rasi
+}
 
 # Ask for confirmation
-function confirm_exit
+confirm_exit() {
 	echo -e "$yes\n$no" | confirm_cmd
-end
+}
 
-function run_rofi
-	echo -e "$reboot\n$shutdown" | rofi_cmd
-end
+# Pass variables to rofi dmenu
+run_rofi() {
+	echo -e "$shutdown\n$reboot\n$lock\n$suspend\n$logout" | rofi_cmd
+}
 
 # Execute Command
-function run_cmd -a command
-	set selected $(confirm_exit)
-	switch $selected
-        case $yes
-	        switch $command
-                case shutdown
-			        systemctl poweroff
-                case reboot
-			        shutdown -r now
-            end
-        case "*"
-		    exit 0
-    end
-end
+run_cmd() {
+	selected="$(confirm_exit)"
+	if [[ "$selected" == "$yes" ]]; then
+		if [[ $1 == '--shutdown' ]]; then
+			systemctl poweroff
+		elif [[ $1 == '--reboot' ]]; then
+			systemctl reboot
+		elif [[ $1 == '--suspend' ]]; then
+			mpc -q pause
+			amixer set Master mute
+			systemctl suspend
+		fi
+	else
+		exit 0
+	fi
+}
 
-set chosen $(run_rofi)
-switch $chosen
-    case $shutdown
-        run_cmd shutdown
-    case $reboot
-        run_cmd reboot
-end
+# Actions
+chosen="$(run_rofi)"
+case ${chosen} in
+    $shutdown)
+		run_cmd --shutdown
+        ;;
+    $reboot)
+		run_cmd --reboot
+        ;;
+    $lock)
+		hyprlock
+        ;;
+    $suspend)
+		run_cmd --suspend
+		hyprlock
+        ;;
+    $logout)
+		hyprctl dispatch exit
+        ;;
+esac
